refactor(core): split App hub orchestration into hooks and components
